# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para sistema de reservas de hotel",
        "title": "API de Reservas",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/reservas": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Devuelve la lista completa de reservas",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reservas"
                ],
                "summary": "Obtener todas las reservas",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Registra una nueva reserva en el sistema",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reservas"
                ],
                "summary": "Crear reserva",
                "parameters": [
                    {
                        "description": "Datos de la reserva",
                        "name": "reserva",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.registerReservaRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/reservas/cliente/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Busca reservas por ID de cliente",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reservas"
                ],
                "summary": "Obtener reservas por cliente",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del cliente",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/reservas/recepcionista/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Busca reservas por ID de recepcionista",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reservas"
                ],
                "summary": "Obtener reservas por recepcionista",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del recepcionista",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/reservas/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Busca una reserva por su ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reservas"
                ],
                "summary": "Obtener reserva por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la reserva",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Actualiza los datos de una reserva existente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reservas"
                ],
                "summary": "Actualizar reserva",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la reserva",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos a actualizar",
                        "name": "reserva",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.updateReservaRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cambia el estado de la reserva en vez de borrarla físicamente",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reservas"
                ],
                "summary": "Eliminar reserva (soft delete)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la reserva",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.registerReservaRequest": {
            "type": "object",
            "required": [
                "cantidadNoches",
                "estadoReserva",
                "fechaReserva",
                "idCliente",
                "idRecepcionista"
            ],
            "properties": {
                "cantidadNoches": {
                    "type": "integer"
                },
                "estadoReserva": {
                    "type": "string"
                },
                "fechaReserva": {
                    "type": "string"
                },
                "idCliente": {
                    "type": "string"
                },
                "idRecepcionista": {
                    "type": "string"
                }
            }
        },
...
        "handlers.updateReservaRequest": {
            "type": "object",
            "required": [
                "cantidadNoches",
                "estadoReserva",
                "fechaReserva",
                "idCliente",
                "idRecepcionista"
            ],
            "properties": {
                "cantidadNoches": {
                    "type": "integer"
                },
                "estado": {
                    "type": "integer"
                },
                "estadoReserva": {
                    "type": "string"
                },
                "fechaReserva": {
                    "type": "string"
                },
                "idCliente": {
                    "type": "string"
                },
                "idRecepcionista": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Escribe \"Bearer \" + tu token JWT. Ejemplo: \"Bearer eyJhbGciOiJIUzI1NiIs...\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}